[MOV] test_mail: move ticket-like models into file
Test models file in test_mail becomes quite large. Better split
functional-like models from models intended to be used for some
corner case or that are feature-specific.
