+++ addons/test_mail/models/__init__.py
@@ -1,4 +1,5 @@
 from . import mail_test_access
+from . import mail_test_ticket
 from . import test_mail_corner_case_models
 from . import test_mail_feature_models
 from . import test_mail_models
